+++ site/dat/docs/Changelog.md
@@ -10,6 +10,7 @@
  - fix Locust crash when `concurrency` isn't specified
  - support scenario-level `headers` in Locust script generator
  - add exception throw to tools downloader
+ - add Mocha autoinstallation for Selenium tests
 
 ## 1.7.0 <sup>2 oct 2016</sup>
  - add RSpec tests runner for Selenium